requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
arxiv==2.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cohere==4.32
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
exceptiongroup==1.1.3
fastapi==0.104.0
fastavro==1.8.2
feedparser==6.0.10
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.1
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.323
langsmith==0.0.52
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openai==0.28.1
orjson==3.9.9
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.24.4
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
pypdf==3.16.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.6.3
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
